ai-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
openai==1.12.0
pydantic==2.6.0
python-dotenv==1.0.0
transformers==4.37.0
torch==2.2.0
sentence-transformers==2.3.1
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.2.0
